Remove OIDC secrets from config template and use BASE_URL instead of CUSTOM_HOSTNAME (#13)
OIDC variables APPLICANT_OIDC_CLIENT_ID and APPLICANT_OIDC_CLIENT_SECRET will be stored in secret manager and shouldn't be provided in the config file. Otherwise they will be stored in the terraform state object which generally isn't considered very secure.

Also replaces CUSTOM_HOSTNAME with BASE_URL as the latter is more widely used in the codebase. CUSTOM_HOSTNAME was moved to the azure section.