Add --force-unlock flag
This adds a --force-unlock flag to the actions, which is passed along to the run script via the -u flag. This will tell the system to run the 'terraform force-unlock' command before doing an apply in order to fix a deployment with the Terraform lock left in the locked state, usually by Ctrl+Cing out of an apply.