Add --fix-digest flag
This flag is for fixing the Terraform state when the Terraform lock has been previously force-unlocked, but two applies were happening at once, resulting in the digest value in DynamoDB for the lock file being an incorrect value. The error message for this gives the value it should be, so --fix-digest allows us to go in and fix that value without having to do it in the AWS console manually.